Move 'template' and 'modify_json' to helpers.rb
